LCh needs at least as much accuracy as HSL
Unfortunately, this breaks the `to_lch_string` test.  The LCh colorspace
is much larger and more precise than the HSL colorspace used to
represent values in the `Color` class.  Even though the test case uses
whole number values for the LCh components when converting into a
`Color`, the closest HSL representation of that color does not produce
whole number results when converted back to LCh.  The numbers are close
enough that the LCh values match when rounding to a whole number, but
not when rounding to one decimal place.