Update test to match the LCh->HSL->LCh round trip
There is some loss of precision when converting an LCh value to an HSL
`Color` instance, so the round trip conversion back to LCh does not
produce an exact match for the original LCh component values.  Precise
round trip conversions are not possible without a fundamental change to
the internal representation of a `Color` object, so we just need to
account for the inaccuracy in the expected result.