Implement compact format for color block output
When displaying many colors in a single command, the standard color
block format limits how many colors can be seen at the same time in the
terminal window.  The compact format cuts the height of each color
block in half by removing the list of similar colors so that more colors
can fit into the height of the terminal window.