Add `luv-lightness` format type
The CIELUV lightness is identical to the CIELAB lightness except for
very small lightness values, but it does use a slightly different
transform for values that are less than the delta-cubed threshold so it
requires a separate format for completeness.  This also renames the
`lch-lightness` format to `lab-lightness`.  CIELAB is a direct analog
to CIELUV, and CIELCh is just a polar transformation of the Cartesian
CIELAB coordinate space.