Convert and format CIELCh(uv) colors
This also adjusts some of the CIELCh(ab) format names for consistency.
The names `lab-lightness` and `luv-lightness` are used for the lightness
values of the CIELAB and CIELUV color spaces.  The CIELCh(ab) and
CIELCh(uv) color spaces are just polar transformations of the ab and uv
coordinates respectively, and use the same lightness values.  To tie
the chroma and hue values to the lightness value, and to avoid any
confusion between the LCh variants, we now use the names `lab-chroma`
and `lab-hue` along with the new `luv-chroma` and `luv-hue` CIELCh(uv)
equivalents.