Decouple the inner details of `Color` storage
The `Color` type currently represents colors using HSL coordinates.
Several of the library functions directly access the HSL properties of
a `Color`, which creates a tight coupling to the internal `Color`
storage format and makes it difficult to change the representation.

If a function naturally operates in the HSL space, we can decouple it
from the internal `Color` storage format by using the `HSLA::from`
conversion function to access the HSL components.  Note that the
conversion functions form the encapsulation layer for the internal
color representation, so we don't try to fully decouple them.  We want
to minimize rounding errors in the conversions by choosing formulas that
convert to or from the internal representation as directly as possible,
and this creates a natural coupling to the internal representation.