Don't try to preserve hue when converting to gray
Preserving the hue angle of a converted gray restricts us to remaining
entirely in the original HSL representation of the color.  In order to
change to a color format that can more accurately represent a broader
range of colors, we must first break the dependence on the HSL format
by always representing pure grays with the conventional 0-degree hue.