Don't try to preserve hue when converting to gray
Cartesian color coordinate systems do not have multiple representations
for the same gray shade, and different polar coordinate systems have
different hue values for the same color, so it is impossible to preserve
the hue angle across format conversions.  The CSS Color Level 4 spec
defines that a pure gray has no hue (a 'none' value), but we don't
currently support that, so we just use the conventional 0-degree hue.