Switch to storing colors as unbounded sRGB
If we use unbounded floating-point RGB components as the common format
for a color, we can accurately represent any possible color, including
colors that would fall outside the normal sRGB gamut.  We only clamp
out-of-range values when converting to a format such as 8-bit sRGB or
HSL that can only represent in-gamut colors.