Improve precision for XYZ to RGB conversions
The sRGB-XYZ transformation and inversion matrices in the original IEC
61966-2-1:1999 sRGB standard were both specified to only 4 decimal
places.  With these matrices, a round trip sRGB-XYZ-sRGB conversion is
only accurate to 3 decimal places.  In 2003, Amendment 1 to the sRGB
standard replaced the original inversion matrix with one specified to
7 decimal places, which increases the round trip accuracy to about 6
decimal places.

Using the white point and RGB reference primaries defined in the sRGB
standard, the XYZ matrix equations can be solved using exact rational
fractions.  The resulting fractions can then be calculated to as many
decimal places as desired.  The example code attached to the CSS Color
Level 4 draft specification takes this approach, and presents the values
to 17 decimal places.  Using these matrices, we can increase the round
trip conversion accuracy to 15 decimal places.  This is important
because it also improves the precision for all of the perceptual CIE
color spaces that derive from XYZ.