Output higher precision for format conversions
Higher precision in the conversion calculations isn't useful unless we
show more precision in the formatted output.  Color values now include
more decimal places in the output, but the CSS serialization rules are
used so trailing zeros are dropped rather than showing all values with
a fixed number of decimal places.  Values are also rounded to the
maximum number of places to be displayed instead of just truncating the
output.