Use high-precision format when piping output
Multi-step calculations require piping color values from one `pastel`
command to another.  To avoid introducing too much error between steps,
we need to use a much higher precision for the output string when piping
a color value to another command.  But, it must be a format that is
recognized as a valid input format by the parser, and should correspond
to the internal storage format so that an extra conversion is not
required.

The format used is the CSS Color Level 4 serialization format for RGB
values.  The [0.0, 1.0] range for the RGB channel values is scaled to
[0.0, 255.0], but out-of-range values are not clamped.  The alpha
channel value is written as a floating point value in the [0.0, 1.0]
range (not as a percentage), and is omitted if it is exactly 1.0.  All
values are output with up to 12 decimal places.