Switch to storing colors in CIE XYZ format
The CIE XYZ color space supports precise floating point channel values
and out-of-gamut colors, and is a common conversion point for many of
the transformations between other color spaces.

For maximum precision in sRGB conversions and calculations, we also use
the floating point format wherever possible and allow out-of-gamut color
values (less than 0 or greater than 1).  Values are only clamped when
converting to a format such as 8-bit sRGB or HSL that can only represent
in-gamut colors.