Improve handling of more precise HSL conversions
When working with precise floating point numbers, rounding error in the
binary fraction representation can play havoc with equality checks.
When we have chroma values that are very close to but not exactly equal
to zero, very small changes in value can cause large swings in the hue
angle as well as incorrect saturation values.  To compensate for this,
we need to use an approximate comparison method.  In this case, checking
for zero equality to 15 decimal places seems to give enough margin to
produce the desired 0° hue angle and 0% saturation for pure grays.

References:
- https://docs.oracle.com/cd/E19957-01/806-3568/ncg_goldberg.html
- https://randomascii.wordpress.com/2012/02/25/comparing-floating-point-numbers-2012-edition/