Add `set` properties for HWB components
Note that HSL, HSV, and HWB all share the same hue value, so setting any
of the `hsl-hue`, `hsv-hue`, or `hwb-hue` properties will have the same
effect.  However, users who are not familiar with the mathematical
derivation of these color representations might not be aware of this
relationship.  The specific `hwb-hue` property makes it obvious that
the value applies to HWB colors.