Use `hsla()` for HSL colors with an alpha value
CSS Color 4 specifies that all RGB colors specified as `rgb()`, `hsl()`,
`hwb()`, or as a named color are serialized as `rgb()` (or `rgba()`).
However, HSL colors can be specified with `hsl()` or `hsla()` which
mirror the `rgb()` and `rgba()` functions, so we can apply the same
rules for formatting HSL colors.