Use percentages for RGB float output
CSS Color 4 specifies that the RGB channel values for the `rgb()` color
function can be specified as either a percentage in the range [0, 100]
or a number in the range [0, 255], and this is also how `rgb()` strings
are parsed by `pastel`.  When formatting a color as an `rgb()` output,
it is preferable to use percentages instead of raw floating point values
in the [0.0, 1.0] range so that the resulting output is a parsable
string.

- https://www.w3.org/TR/css-color-4/#rgb-functions