Format _negative zero_ without the negative sign
Rust supports the concept of a "negative zero", which is a floating
point zero value where the sign is negative.  One example of where this
can occur is when a small amount of error in a calculation results in a
value that is just slightly less than zero.  When this tiny negative
value rounds to zero, it is represented as negative zero.

When formatting a negative zero, Rust will represent it as "-0", which
doesn't really make sense in a color representation.  To avoid this, we
need to check for this special condition when rounding, and convert the
"negative zero" to the normal "positive zero".