Use `device-cmyk()` function for CMYK colors
CSS Color 4 specifies that the `device-cmyk()` function should be used
for serializing CMYK colors, using numeric values in the [0, 1] range.
However, in almost any other context, CMYK color values are presented
as either a percentage or a number on a [0, 100] scale.  CSS Color 4
allows percentages for defining colors with the `device-cmyk()`
function, so we choose to use that format instead of the specified
[0, 1] numeric values.

- https://www.w3.org/TR/css-color-4/#serializing-device-cmyk-values
- https://www.w3.org/TR/css-color-4/#device-cmyk