Split out modern `rgb` and `hsl` format tests
For the `rgb()` and `hsl()` functions, CSS-Color-4 defines both a
"legacy" syntax and a "modern" syntax.  Both of these were _mostly_
handled by the original parser.

This splits the checks for the modern syntax out into separate test
functions so that it is easier to review the coverage, then fills in
some missing tests and updates some incorrect expected results to
properly test the full CSS-Color-4 syntax.  Note that this results in
a few failing tests, mostly resulting from incomplete handling of the
CSS alpha channel syntax.