Update `rgb` and `hsl` tests for lenient modern parsing
The "modern" CSS syntax requires a '/' separator before the alpha value,
but we want to make the slash optional and allow all arguments to be
separated by spaces.  Therefore, we need to check that this format
produces a correct color rather than a missing output.

This also renames the `rgb` and `hsl`test functions for better naming
consistency.