Update the version number
This fork should be "mostly" compatible with the original `pastel`
utility, but it is not a goal to guarantee 100% compatibility.  This
version of the utility has evolved new functionality in several areas,
including new color spaces, additional input and output formats, and
more precise color conversions.  Therefore, it seems warranted at this
point to at least update the original version number.  If this fork
continues to diverge from the original in both features and goals, then
it may eventually become necessary to give the project a different name.