Remove `Format::Spaces` options
The `NoSpaces` option has very little value when using the normative
CSS Color 4 functional notation because the only place where it is valid
to omit the spaces is around the '/' before an optional alpha value.
Most of the usages from the CLI are hard-coded to use the `Spaces`
option anyway.

This greatly simplifies the implementation at the cost of a breaking
change to the public library API.  However, the formats of the returned
strings will be changing anyway, so API compatibility is not a primary
goal here.