Use matrices to simplify conversion algorithms
Many color conversions are defined mathematically using transformation
matrices and matrix multiplication, so using matrices in the conversion
algorithms allows a more direct mapping to the mathematical definitions.
This eliminates a lot of boilerplate code to implement the equivalent
multiplications explicitly in each algorithm, which results in better
readability and simplifies the addition of new conversions.  It also
provides a single point where all matrix multiplications could be
optimized for performance or improved precision, if needed.