Extract HSV color space into a sub-module
Various pieces of the functionality for a given color space are
currently scattered throughout the `lib.rs` file.  This makes it
challenging to find all of those pieces and keep them organized as more
color spaces are added.  Using the relatively simple HSV color space as
a proof-of-concept, this examines what it would look like to group all
of the related functionality for each color space together into a
separate sub-module.