Refactor CSSC4 `color()` function parsing
Originally, all variants of the CSSC4 `color()` function were handled by
a single parse function, and the parsers for specific color spaces were
just sub-parsers used by that function.  This prevented the parsers from
being moved to the appropriate sub-modules because it was not possible
to isolate the parsing of a complete `color()` function string for a
single color space.

This refactoring converts the common portions of the `color()` function
parser into a shared parser function that takes the sub-parsers for a
specific color space as parameters.  This allows parsing of the common
`color()` function syntax to be shared, while also providing complete
and independent parser functions for each color space.