Rename `Rgba` -> `Srgba`
RGB is not a specific color space, but a generic term for any color
specified by R, G, and B component values.  What is actually implemented
in `pastel` currently is the sRGB color space.  Renaming the struct to
`Srgba` makes this more explicit, and also opens the door to implement
additional RGB-based color spaces, such as Display-P3 or Rec. 2020.