Implement `From` traits on owned types
The `From` trait is documented as providing "value-to-value conversions
while consuming the input value," but a `From` conversion from a
borrowed reference to an owned value does not consume the input.  It can
also be quite tricky to get all of the lifetime specifications correct
when borrowed types are used as type arguments in a trait bound.